Cập nhật code làm bài thi
+ Sửa ảnh nền trang liên hệ
+ Sửa lỗi truy vấn nhỏ trong trang làm bài thi
+ Tạo riêng controller cho chức năng làm bài thi
+ Câu truy vấn làm bài thi được sử dụng chung trong model_general (model sử dụng chung cho hầu hết chức năng bên ngoài)
