+++ .gitignore
@@ -1,3 +1,4 @@
 # bỏ qua file để đỡ phải sửa đi sửa lại thông tin đăng nhập vô database
 database/config.php
 View/.htaccess
+database/config.php
