+++ View/testing.php
@@ -53,5 +53,5 @@
 
         <div style="clear: both"></div>
         <br>
-    </div>
+    </div><br><br>
 </div>
