+++ database/config.php
@@ -1,7 +1,7 @@
 <?php
 	define("sql_server","localhost");
     define("sql_user","root");
-    define("sql_password","");
+    define("sql_password","thienvip123");
 	define("sql_database","toeicdb");
 	define("BASE_URL","http://localhost/ToeicThi/");
 	define("BASE_DIR","/ToeicThi/");
